fix: use bf16 for SDXL training (fp16 causes NaN loss)
- SDXL UNet activations overflow fp16 max (~65504), producing NaN
- VAE latent caching also corrupted with fp16 (54% of cached latents had NaN)
- bf16 shares fp32 exponent range, works correctly
- Also: add training status CLI, preview UI, clippy fixes

+++ python/mods_worker/adapters/train_adapter.py
@@ -111,7 +111,7 @@ def _build_train_block(family: str, params: dict, lora_type: str) -> dict:
     elif family == "sdxl":
         train["train_text_encoder"] = True
         train["noise_scheduler"] = "ddpm"
-        train["dtype"] = "fp16"
+        train["dtype"] = "bf16"  # fp16 causes NaN loss due to overflow in SDXL UNet
         train["max_denoising_steps"] = 1000
         # SDXL was trained with noise_offset 0.0357
         train["noise_offset"] = 0.0357 if is_style else 0.0
